refactor: extract SystemScreenCapturePermissionAccess.swift from ScreenCapturePermissionService.swift
Splits the `SystemScreenCapturePermissionAccess` struct (~52 lines,
the concrete ScreenCapturePermissionAccessing implementation) out of
ScreenCapturePermissionService.swift into its own file.

Byte-identical move of lines 5-56. Both files import AppKit +
CoreGraphics + Foundation.

ScreenCapturePermissionService.swift: 209 → 156 lines. Tests: 667.
